Moved PersistenceInjectionModule to persistance package instead of persistence.storage
Added export metricsService to the module-info

Refactored Dagger Component to receive a Configuration from outside.

Added UT for MediatorInjectionModule
