+++ server/src/main/java/module-info.java
@@ -13,6 +13,7 @@ module com.hedera.block.server {
     exports com.hedera.block.server.metrics;
     exports com.hedera.block.server.data;
     exports com.hedera.block.server.health;
+    exports com.hedera.block.server.persistence;
 
     requires com.hedera.block.stream;
     requires com.google.protobuf;
